+++ axelor-account/src/main/java/com/axelor/apps/account/service/move/MoveServiceImpl.java
@@ -19,13 +19,16 @@ package com.axelor.apps.account.service.move;
 
 import com.axelor.apps.account.db.Account;
 import com.axelor.apps.account.db.AccountConfig;
+import com.axelor.apps.account.db.AnalyticMoveLine;
 import com.axelor.apps.account.db.Invoice;
 import com.axelor.apps.account.db.Journal;
 import com.axelor.apps.account.db.Move;
 import com.axelor.apps.account.db.MoveLine;
 import com.axelor.apps.account.db.Reconcile;
+import com.axelor.apps.account.db.repo.AnalyticMoveLineRepository;
 import com.axelor.apps.account.db.repo.MoveRepository;
 import com.axelor.apps.account.db.repo.ReconcileRepository;
+import com.axelor.apps.account.service.AnalyticMoveLineService;
 import com.axelor.apps.account.service.ReconcileService;
 import com.axelor.apps.account.service.app.AppAccountService;
 import com.axelor.apps.account.service.config.AccountConfigService;
@@ -467,6 +470,19 @@ public class MoveServiceImpl implements MoveService {
       Boolean isDebit = moveLine.getDebit().compareTo(BigDecimal.ZERO) > 0;
 
       MoveLine newMoveLine = generateReverseMoveLine(newMove, moveLine, dateOfReversion, isDebit);
+
+      if (moveLine.getAnalyticDistributionTemplate() != null) {
+        newMoveLine.setAnalyticDistributionTemplate(moveLine.getAnalyticDistributionTemplate());
+
+        List<AnalyticMoveLine> analyticMoveLineList =
+            Beans.get(AnalyticMoveLineService.class)
+                .generateLines(
+                    newMoveLine.getAnalyticDistributionTemplate(),
+                    newMoveLine.getDebit().add(newMoveLine.getCredit()),
+                    AnalyticMoveLineRepository.STATUS_REAL_ACCOUNTING,
+                    move.getDate());
+        newMoveLine.setAnalyticMoveLineList(analyticMoveLineList);
+      }
 
       newMove.addMoveLineListItem(newMoveLine);