+++ CHANGELOG.md
@@ -288,5 +288,6 @@
 - ACCOUNTING MOVE : fix NPE when reverse the move
 - INVOICE : fix for NPE when printing from grid view
 - Advance Import : Resolve ArrayIndexOurOfBound exception.
+- ACCOUNTING : reversion and analytic distribution
 
 [Unreleased 5.2.0]: https://github.com/axelor/axelor-open-suite/compare/dev...wip